Move the tests for deferredutil into their own module.
